renamed app.py to routes.py, separated tests
